// src/app/shared/types/api-types/bire-evolution-dto.interface.ts
export interface BireEvolutionDTO {
  evolutionNumber?: string;
  evolutionRevisionNumber?: string;
  evolutionType?: string;
  evolutionSummary?: string;
  evolutionLevel?: string;
  evolutionCompliance?: string;
  tracked?: boolean;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}