// src/app/shared/types/api-types/bire-sb-dto.interface.ts
export interface BireSbDTO {
  sbNumber?: string;
  sbRevisionNumber?: string;
  sbTitle?: string;
  sbType?: string;
  sbClass?: string;
  sbDocumentUrl?: string;
  sbAssets?: string;
  hoursLimit?: string;
  cyclesLimit?: string;
  limitComplianceDate?: Date;
  sbPublicationDate?: Date;
  sbEffectiveDate?: Date;
  sbCompliance?: string;
  sbComplianceStatus?: string;
  sbComplianceInformation?: string;
  sbSubType?: string;
  sbContext?: string;
  sbComments?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}